Generalize `instance HasLoggingServer (VerbMod ...)`
